nvme: check the PRINFO bit before deciding the host buffer
Git-commit: 57f26d9d09d64e0b31bedffdd0de460215675ca9
 length
References: bsn#19
Patch-mainline: v5.10.14

[ Upstream commit 4d6b1c95b974761c01cbad92321b82232b66d2a2 ]

According to NVMe spec v1.4, section 8.3.1, the PRINFO bit and
the metadata size play a vital role in deteriming the host buffer size.

If PRIFNO bit is set and MS==8, the host doesn't add the metadata buffer,
instead the controller adds it.

Signed-off-by: Revanth Rajashekar <revanth.rajashekar@intel.com>
Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.14-0039-nvme-check-the-PRINFO-bit-before-deciding-the-host-b.patch