mptcp: do not queue excessive data on subflows
Git-commit: ec369c3a337fe075a7bd4da88d163d44c62ccbb1
Message-Id: <ec369c3a337fe075a7bd4da88d163d44c62ccbb1.1637074654.git.geliang.tang@suse.com>
References: bsn#2
Patch-mainline: v5.12-rc1

The current packet scheduler can enqueue up to sndbuf
data on each subflow. If the send buffer is large and
the subflows are not symmetric, this could lead to
suboptimal aggregate bandwidth utilization.

Limit the amount of queued data to the maximum send
window.

Reviewed-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
Signed-off-by: Geliang Tang <geliang.tang@suse.com>

Patch-name: patches.suse/v5.12-rc1-mptcp-do-not-queue-excessive-data-on-subflows.patch